Put converted mp3s into folders by disc.
The player on my phone can't tell that the RHCP collection is a series
of separate discs, so I get to hear the entire 100-song collection when
I'm playing random albums. Maybe it can and I just haven't figured out
how to get it to detect that. I don't know. What I do know is that it
can tell the difference between albums that have the same name if they
are in different folders (why does everyone want to name their album I,
or II, or III, or IV; stupid Romans).

This tells beets to put a disc number in the folder name if there is a
disc in the metadata of the source file. Beets puts disc numbers on
everything that I import, so everything has a disc.